--- a/include/rfl/parsing/Parser_rfl_variant.hpp
+++ b/include/rfl/parsing/Parser_rfl_variant.hpp
@@ -68,6 +68,7 @@ class Parser<R, W, rfl::Variant<AlternativeTypes...>, ProcessorsType> {
     } else {
       std::optional<rfl::Variant<AlternativeTypes...>> result;
       std::vector<Error> errors;
+      errors.reserve(sizeof...(AlternativeTypes));
       read_variant(
           _r, _var, &result, &errors,
           std::make_integer_sequence<int, sizeof...(AlternativeTypes)>());
@@ -187,7 +188,7 @@ class Parser<R, W, rfl::Variant<AlternativeTypes...>, ProcessorsType> {
       if (res) {
         *_result = std::move(*res);
       } else {
-        _errors->emplace_back(res.error());
+        _errors->emplace_back(std::move(res.error()));
       }
     }
   }