@liuzicheng1987 2025-08-20
Ok. Could you try to just set err = std::nullopt when you first declare it?