@GregTheMadMonk 2025-08-20
It didn't help, sadly, for me at least. Same with replacing `if (err)` with `if (err.has_value())`:
<img width="1279" height="1374" alt="image" src="https://github.com/user-attachments/assets/dc04c2e3-11aa-496f-9fa3-23eb46a4e50f" />