@liuzicheng1987 2025-08-24
@GregTheMadMonk , this should fix the issue. We'll wait for the tests to run through and then this is ready to be merged.

Thank you for the contribution!