@GregTheMadMonk 2025-08-25
Thank you!

How did you come to this fix? Or was that the warning system in GCC has actually misfired and you had to do it by trial and error?